% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Writing materials.R
\name{add_heading_after_identical_brackets_pair}
\alias{add_heading_after_identical_brackets_pair}
\title{Add a Secondary Title After Identical Brackets Pair}
\usage{
add_heading_after_identical_brackets_pair(text)
}
\arguments{
\item{text}{The text in which to identify and enhance pairs of identical brackets.}
}
\value{
Returns the text with added secondary titles after each pair of identical brackets.
}
\description{
For pairs of identical bracketed contents, this function adds a secondary title after each pair to improve document structure and readability.
}
\examples{
add_heading_after_identical_brackets_pair("Text before.【Same】Text in between.【Same】Text after.")
}
