% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DXMeSH.R
\name{dx_mesh}
\alias{dx_mesh}
\alias{get_mesh}
\title{DXMeSH: Create Comprehensive and Accurate Search Queries Based on MeSH}
\usage{
get_mesh(csv_file)
}
\arguments{
\item{csv_file}{A string that specifies the path of the CSV file containing MeSH terms and their corresponding entries.}
}
\value{
The function does not return a value but creates a "query.txt" file containing the final search query.
}
\description{
This function reads a CSV file of MeSH terms and their corresponding entries,
creates a search query for each keyword and combines all queries with an AND operation.
The final search query is saved in a text file named "query.txt".
}
\examples{
dx_mesh("DXMeSH.csv")
}
