% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Writing materials.R
\name{get_Writing_materials}
\alias{get_Writing_materials}
\title{Main function to process the document and convert its format}
\usage{
get_Writing_materials(input_file)
}
\arguments{
\item{input_file}{The path to the input file which should be processed.}
}
\value{
This function does not return a value but saves processed output in specified formats.
}
\description{
This function processes an input document, applies text processing functions, and saves the results in markdown and Word formats.
}
\examples{
get_Writing_materials("引言-1.docx")
}
