% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Writing materials.R
\name{mark_even_occurrences}
\alias{mark_even_occurrences}
\title{Add Mark to Second Bracket of Every Pair with Same Content}
\usage{
mark_even_occurrences(text, mark = "### ")
}
\arguments{
\item{text}{The text in which to mark the second bracket of each pair with identical content.}
}
\value{
Returns the modified text with marks added before the second bracket of every identical pair.
}
\description{
This function adds a specific mark (###) before the second bracket of every pair that has identical content within the text, aiding in the emphasis or distinction of repeated elements.
}
\examples{
mark_even_occurrences("Some introductory text 【Bracket Content】 followed by more text and then 【Bracket Content】 again.")
}
