% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Writing materials.R
\name{read_markdown}
\alias{read_markdown}
\title{Read the Markdown file}
\usage{
read_markdown(file_path)
}
\arguments{
\item{file_path}{The path to the markdown file.}
}
\value{
Returns the content of the markdown file as a single string.
}
\description{
This function reads the content of a markdown file and returns it as a single string.
}
\examples{
read_markdown("example.md")
}
