% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Writing materials.R
\name{substr_replace}
\alias{substr_replace}
\title{Auxiliary function: Inserts text at a specific position in the string}
\arguments{
\item{string}{The original string.}

\item{replacement}{The string to be inserted.}

\item{start}{The starting position for the replacement.}

\item{stop}{The ending position for the replacement.}
}
\value{
Returns a modified string with the replacement text inserted.
}
\description{
This function takes a string and inserts a replacement string between the start and stop positions.
}
\examples{
substr_replace("The quick brown fox", "happy ", 10, 10)
}
