% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Writing materials.R
\name{write_markdown}
\alias{write_markdown}
\title{Write back to Markdown file}
\usage{
write_markdown(text, file_path)
}
\arguments{
\item{text}{The text to be written to the file.}

\item{file_path}{The path where the markdown file will be saved.}
}
\value{
This function does not return a value.
}
\description{
This function writes a string of text back to a markdown file.
}
\examples{
write_markdown("# New Markdown Content", "example.md")
}
